+++ CHANGES.rst
@@ -33,6 +33,10 @@ Unreleased
     stranded there and not combined into the totals, as described in `issue
     2025`_.  This is now fixed.
 
+  - On Windows (or maybe only some Windows?) the patch would fail with a
+    ``ModuleNotFound`` error trying to import coverage.  This is now fixed,
+    closing `issue 2022`_.
+
 - Fix: really close all SQLite databases, even in-memory ones. Closes `issue
   2017`_.
 
@@ -42,6 +46,7 @@ Unreleased
 .. _issue 2007: https://github.com/nedbat/coveragepy/issues/2007
 .. _issue 2017: https://github.com/nedbat/coveragepy/issues/2017
 .. _pull 2018: https://github.com/nedbat/coveragepy/pull/2018
+.. _issue 2022: https://github.com/nedbat/coveragepy/issues/2022
 .. _issue 2024: https://github.com/nedbat/coveragepy/issues/2024
 .. _issue 2025: https://github.com/nedbat/coveragepy/issues/2025